+++ NEWS.md
@@ -12,6 +12,10 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.16.0
 
+Released on CRAN 202X-XX-XX
+* Added parameter 'VariantKey' to tidy functions to provide a unique identifier for each variant.
+* vcfR::vcf_field_names() now uses `sub("^[^=]+=", "", myValue)` to remove key from meta key-pairs.
+
 
 # vcfR 1.15.0
 Released on CRAN 2023-12-07